Undo/Redo reimplemented; big Flow and NodeInstance cleanup
A big one.

The new system now works with redo/undo for all NodeInstances - the handling of the DrawingObjects still needs adjustment in the Flow, but it should be very easy from now on.
Cleaned up NodeInstance class and Flow class.
The whole selection processes as well as undo/redo-operations are way nicer now. The undo/redo system now does not rely anymore on indices (!!) but on actual references.