Implemented the most part of the new NI layout system
Removing ports is still a problem, but everything else works for now. That might hold me up for a while though, some really strange QGraphicsLinearLayout behaviour is going on.

Also added an option to choose a mode for algorithm synchronization which might becaome a very interesting feature in the future.