Added retain method to keep references intact when editing methods
For the usage of method objects (e.g. when connecting a slot to a signal) there is now a method m() available to enture that in the case of an edited method, the newly edited one is going to be used instead of the old one.
I also added the import of this method to all existing packages.