removed widget_type argument in all input widget related methods and containers
Instead of widget_type, the widget_name is going to be used directly, also for std input widgets.
Not using an input widget is now indicated by setting (or leaving) widget_name to None.
I adapted all nodes and the save&load procedures of Ryven as well as the NodeManager.