bubble sort example and some new nodes
Additionally, I discovered that there is an issue when using the forward propagating algorithm mode in active flows regarding the autonomous change of values of variables. Since I cannot see a simple and performant solution to this, I will require the user to always choose the appropriate algorithm mode for the script he is trying to execute. I hate that but for now it seems necessary.