personalized algorithm and sync modes for flows
Every flow now has his personal settings are for algorithm mode and viewport update synchronization mode since choosing an appropriate algorithm mode seems to be a necessity with a lot of algorithmic flows. These settings get saved and loaded to and from project files, backward compatibility is given. I am going to add an explanation of this situation to the docs.