updated links and code formatting
The first link I changed had led to a 404, the second one now goes to the welcome page of the ryvencore-qt docs instead of some chapter of it.
In some indented blocks, the code didn't appear inside the code-field, but below it, and was therefore not formatted as code. Changed that too, I hope for the better.