single-sourcing version in setup.cfg
I really don't enjoy the stuff flying around. Package metadata should be in one, and only one place, providing a single source of truth. With the current build setup that's setup.cfg. I wrote a small utility function that manually reads the version from setup.cfg when the file exists. If not - that is when Ryven is installed as package - the version is provided by importlib.metadata.