Per package menu
This is a quick update in order to provide better package experience in the editor. Though ultimately I believe this system needs to be remade and possibly added to ryvencore.

Packages are now added in a list -per name. There's also a trick to include "subpackages". Packages can be named as {root}.{subpackage}. There are no subpackages of subpackages, it only works for one level of grouping. By clicking on the last node of a package, the corresponding nodes are shown for Ryven.