Implemented sub-packaging
-There is no notion of a sub-package as in a class, it's still the same NodesPackage. 'export_nodes' now builds a dict of str to (nodes, datas). Keys are a combination of the package and subpackage names. Subpackages can be in a folder, where the main nodes.py can be called anything and simply imports the sub-packages.

On each sub-package, 'export_sub_package' is called to extend the package name. A name can be given as a function parameter. If not, it uses the module name if it isn't named "nodes.py", in which case it uses the folder name. Sub-packages can exist in a subfolder or the main nodes.py folder, in which case the name should be different.