Select / Delete Utility for Connections
-Added the ability to select and delete connections.
-Implemented the paint method for ConnectionItem so that it doesn't draw its bounding rect when selected.

Just a notice, FlowView and FlowCommands are in the same module but depend on each other. We cannot import FlowView to FlowCommands for better type hinting due to partial initialization caused by circular imports. Perhaps the code in FlowCommands should be added in FlowView, rather than being kept as a separate module?